fix CR #1317079

+++ libs/dimg/filters/dimgthreadedfilter.cpp
@@ -6,7 +6,7 @@
  * Date        : 2005-05-25
  * Description : threaded image filter class.
  *
- * Copyright (C) 2005-2014 by Gilles Caulier <caulier dot gilles at gmail dot com>
+ * Copyright (C) 2005-2015 by Gilles Caulier <caulier dot gilles at gmail dot com>
  * Copyright (C) 2007-2012 by Marcel Wiesweg <marcel dot wiesweg at gmx dot de>
  *
  * This program is free software; you can redistribute it
@@ -276,7 +276,7 @@ QString DImgThreadedFilter::readParametersError(const FilterAction&) const
 QList<int> DImgThreadedFilter::multithreadedSteps(int stop, int start) const
 {
     uint  nbCore = QThreadPool::globalInstance()->maxThreadCount();
-    float step   = (stop - start) / nbCore;
+    float step   = ((float)stop - (float)start) / (float)nbCore;
     QList<int> vals;
 
     vals << start;
